requirements.txt
numpy
pandas
python-dotenv
mysql-connector-python
pymysql
scikit-learn
catboost
xgboost
Flask
dagshub
mlflow
# -e .